// FRONTEND/chatbot-admin-panel/src/services/api/schemas/authorizedMenuResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { AuthorizedMenuResponseIconName } from './authorizedMenuResponseIconName';
import type { AuthorizedMenuResponseParentId } from './authorizedMenuResponseParentId';

export interface AuthorizedMenuResponse {
  id: number;
  name: string;
  frontend_route: string;
  icon_name: AuthorizedMenuResponseIconName;
  parent_id: AuthorizedMenuResponseParentId;
  display_order: number;
}